# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RoastMySite API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface ReportInput {
  url: string;
}

export interface ScoreCategory {
  name: string;
  score: number;
  maxScore: number;
  /** @nullable */
  description?: string | null;
}

/**
 * @nullable
 */
export type RoastSectionSeverity =
  | (typeof RoastSectionSeverity)[keyof typeof RoastSectionSeverity]
  | null;

export const RoastSectionSeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;

export interface RoastSection {
  title: string;
  content: string;
  /** @nullable */
  severity?: RoastSectionSeverity;
}

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus];

export const ReportStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface Report {
  id: number;
  userId: string;
  url: string;
  status: ReportStatus;
  /** @nullable */
  screenshotUrl?: string | null;
  /** @nullable */
  pageTitle?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  scoreCategories?: ScoreCategory[] | null;
  /** @nullable */
  firstImpression?: string | null;
  /** @nullable */
  roastSummary?: string | null;
  /** @nullable */
  sections?: RoastSection[] | null;
  /** @nullable */
  rewrittenHeadline?: string | null;
  /** @nullable */
  rewrittenCta?: string | null;
  /** @nullable */
  quickWins?: string[] | null;
  /** @nullable */
  shareSlug?: string | null;
  isPublic?: boolean;
  isFavorite: boolean;
  /** @nullable */
  lighthouseScore?: number | null;
  /** @nullable */
  performanceScore?: number | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ShareLink {
  shareSlug: string;
  shareUrl: string;
}

export interface UserStats {
  totalReports: number;
  todayReports: number;
  dailyLimit: number;
  favoriteReports: number;
  /** @nullable */
  avgScore: number | null;
  isPro: boolean;
  credits: number;
}

export interface User {
  id: number;
  clerkId: string;
  /** @nullable */
  email?: string | null;
  isPro: boolean;
  credits: number;
  /** @nullable */
  stripeCustomerId?: string | null;
  /** @nullable */
  stripeSubscriptionId?: string | null;
  dailyRoastsUsed: number;
  /** @nullable */
  dailyRoastsResetAt?: string | null;
  createdAt: string;
}

export interface Plan {
  id: string;
  name: string;
  price: number;
  interval: string;
  /** @nullable */
  stripePriceId?: string | null;
  features: string[];
}

export interface CheckoutInput {
  priceId: string;
}

export interface CheckoutSession {
  url: string;
}

export interface PortalSession {
  url: string;
}

export interface GrantCreditsInput {
  /** @minimum 1 */
  amount: number;
  /** @nullable */
  note?: string | null;
}

export interface CreditGrantResult {
  clerkId: string;
  credits: number;
  added: number;
}

export interface AdminUser {
  id: number;
  clerkId: string;
  /** @nullable */
  email?: string | null;
  isPro: boolean;
  credits: number;
  createdAt: string;
}

export type ListReportsParams = {
  limit?: number;
  offset?: number;
};

export type ListAdminUsersParams = {
  limit?: number;
  offset?: number;
};