# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RoastMySite API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reports
    description: Website roast reports
  - name: users
    description: User management
  - name: billing
    description: Subscription and billing
  - name: admin
    description: Admin operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List user's roast reports
      description: Returns paginated list of roast reports for the authenticated user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Submit a URL for roasting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "202":
          description: Report accepted, processing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Invalid URL or limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: No credits remaining
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a specific roast report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a roast report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Report deleted
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /reports/{id}/favorite:
    patch:
      operationId: toggleFavoriteReport
      tags: [reports]
      summary: Toggle favorite status of a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /reports/{id}/share:
    post:
      operationId: shareReport
      tags: [reports]
      summary: Generate a public share link for a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Share link details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLink"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /reports/public/{shareSlug}:
    get:
      operationId: getPublicReport
      tags: [reports]
      summary: Get a publicly shared roast report (no auth required)
      parameters:
        - name: shareSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /reports/stats:
    get:
      operationId: getReportStats
      tags: [reports]
      summary: Get usage stats and report summary for the current user
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /billing/plans:
    get:
      operationId: getBillingPlans
      tags: [billing]
      summary: Get available subscription plans
      responses:
        "200":
          description: Available plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"

  /billing/checkout:
    post:
      operationId: createCheckoutSession
      tags: [billing]
      summary: Create Stripe checkout session for upgrading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"

  /billing/portal:
    post:
      operationId: createBillingPortalSession
      tags: [billing]
      summary: Create Stripe billing portal session
      responses:
        "200":
          description: Portal session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalSession"

  /admin/users/{clerkId}/credits:
    post:
      operationId: grantCredits
      tags: [admin]
      summary: Manually grant credits to a user (admin only)
      parameters:
        - name: clerkId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantCreditsInput"
      responses:
        "200":
          description: Updated user credits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditGrantResult"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin only)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    ReportInput:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri

    ScoreCategory:
      type: object
      required:
        - name
        - score
        - maxScore
      properties:
        name:
          type: string
        score:
          type: number
        maxScore:
          type: number
        description:
          type: ["string", "null"]

    RoastSection:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
        content:
          type: string
        severity:
          type: ["string", "null"]
          enum: ["critical", "warning", "info", null]

    Report:
      type: object
      required:
        - id
        - userId
        - url
        - status
        - isFavorite
        - createdAt
      properties:
        id:
          type: integer
        userId:
          type: string
        url:
          type: string
        status:
          type: string
          enum: ["pending", "processing", "completed", "failed"]
        screenshotUrl:
          type: ["string", "null"]
        pageTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        overallScore:
          type: ["number", "null"]
        scoreCategories:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/ScoreCategory"
        firstImpression:
          type: ["string", "null"]
        roastSummary:
          type: ["string", "null"]
        sections:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/RoastSection"
        rewrittenHeadline:
          type: ["string", "null"]
        rewrittenCta:
          type: ["string", "null"]
        quickWins:
          type: ["array", "null"]
          items:
            type: string
        shareSlug:
          type: ["string", "null"]
        isPublic:
          type: boolean
          default: false
        isFavorite:
          type: boolean
          default: false
        lighthouseScore:
          type: ["number", "null"]
        performanceScore:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time

    ShareLink:
      type: object
      required:
        - shareSlug
        - shareUrl
      properties:
        shareSlug:
          type: string
        shareUrl:
          type: string

    UserStats:
      type: object
      required:
        - totalReports
        - todayReports
        - dailyLimit
        - favoriteReports
        - avgScore
        - isPro
        - credits
      properties:
        totalReports:
          type: integer
        todayReports:
          type: integer
        dailyLimit:
          type: integer
        favoriteReports:
          type: integer
        avgScore:
          type: ["number", "null"]
        isPro:
          type: boolean
        credits:
          type: integer

    User:
      type: object
      required:
        - id
        - clerkId
        - isPro
        - credits
        - dailyRoastsUsed
        - createdAt
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: ["string", "null"]
        isPro:
          type: boolean
        credits:
          type: integer
        stripeCustomerId:
          type: ["string", "null"]
        stripeSubscriptionId:
          type: ["string", "null"]
        dailyRoastsUsed:
          type: integer
        dailyRoastsResetAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time

    Plan:
      type: object
      required:
        - id
        - name
        - price
        - interval
        - features
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        interval:
          type: string
        stripePriceId:
          type: ["string", "null"]
        features:
          type: array
          items:
            type: string

    CheckoutInput:
      type: object
      required:
        - priceId
      properties:
        priceId:
          type: string

    CheckoutSession:
      type: object
      required:
        - url
      properties:
        url:
          type: string

    PortalSession:
      type: object
      required:
        - url
      properties:
        url:
          type: string

    GrantCreditsInput:
      type: object
      required:
        - amount
      properties:
        amount:
          type: integer
          minimum: 1
        note:
          type: ["string", "null"]

    CreditGrantResult:
      type: object
      required:
        - clerkId
        - credits
        - added
      properties:
        clerkId:
          type: string
        credits:
          type: integer
        added:
          type: integer

    AdminUser:
      type: object
      required:
        - id
        - clerkId
        - isPro
        - credits
        - createdAt
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: ["string", "null"]
        isPro:
          type: boolean
        credits:
          type: integer
        createdAt:
          type: string
          format: date-time