# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RoastMySite API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: number;
  clerkId: string;
  /** @nullable */
  email?: string | null;
  isPro: boolean;
  credits: number;
  createdAt: Date;
}