# lib/api-zod/src/generated/types/creditGrantResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RoastMySite API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreditGrantResult {
  clerkId: string;
  credits: number;
  added: number;
}