# lib/api-zod/src/generated/types/grantCreditsInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RoastMySite API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GrantCreditsInput {
  /** @minimum 1 */
  amount: number;
  /** @nullable */
  note?: string | null;
}